Authcontroller vendor and Admin session and  login ready

+++ DTOs/Pagenation.cs
@@ -7,5 +7,5 @@ public class Pagination
     [FromQuery, Range(1, int.MaxValue)]
     public int Page { get; set; } = 1;
     [FromQuery, Range(1, int.MaxValue)]
-    public int Limit { get; set; } = 10;
+    public ushort Limit { get; set; } = 10;
 }
